requirements.txt
biopython==1.76
docopt==0.6.2